fix: sidebar layout, updates layout, break files
